feat(tree): add No.99 & No.669 & No.208 solutions

+++ LeetCodeClang/src/tree/p637_average_of_levels_in_binary_tree/average_of_levels_in_binary_tree.cpp
@@ -15,7 +15,7 @@ std::vector<double> Solution637::averageOfLevels(TreeNode* root)
 
 	while (!level_nodes.empty())
 	{
-		int nodes_count = level_nodes.size();
+		int nodes_count = static_cast<int>(level_nodes.size());
 		double sum = 0;
 
 		for (int i = 0; i < nodes_count; ++i)
